pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "callstack"
version = "0.1.0"
description = "Async-first GSM/LTE modem telephony framework for Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "pyserial-asyncio>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
sqlite = [
    "aiosqlite",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
